Make ContainerThreadPool an interface
