+++ config-model/src/main/java/com/yahoo/vespa/model/container/ContainerThreadpool.java
@@ -4,6 +4,7 @@ package com.yahoo.vespa.model.container;
 import com.yahoo.container.bundle.BundleInstantiationSpecification;
 import com.yahoo.container.handler.threadpool.ContainerThreadPool;
 import com.yahoo.container.handler.threadpool.ContainerThreadpoolConfig;
+import com.yahoo.container.handler.threadpool.DefaultContainerThreadpool;
 import com.yahoo.osgi.provider.model.ComponentModel;
 import com.yahoo.text.XML;
 import com.yahoo.vespa.model.container.component.SimpleComponent;
@@ -29,7 +30,7 @@ public class ContainerThreadpool extends SimpleComponent implements ContainerThr
         super(new ComponentModel(
                 BundleInstantiationSpecification.getFromStrings(
                         "threadpool@" + name,
-                        ContainerThreadPool.class.getName(),
+                        DefaultContainerThreadpool.class.getName(),
                         null)));
         this.name = name;
         this.userOptions = userOptions;
