+++ container-core/src/main/java/com/yahoo/container/core/config/testutil/HandlersConfigurerTestWrapper.java
@@ -13,8 +13,6 @@ import com.yahoo.container.core.config.HandlersConfigurerDi;
 import com.yahoo.container.di.CloudSubscriberFactory;
 import com.yahoo.container.di.ComponentDeconstructor;
 import com.yahoo.container.handler.threadpool.ContainerThreadPool;
-import com.yahoo.container.handler.threadpool.ContainerThreadpoolConfig;
-import com.yahoo.jdisc.Metric;
 import com.yahoo.jdisc.handler.RequestHandler;
 import com.yahoo.language.Linguistics;
 import com.yahoo.language.simple.SimpleLinguistics;
@@ -22,9 +20,10 @@ import com.yahoo.language.simple.SimpleLinguistics;
 import java.io.File;
 import java.io.IOException;
 import java.util.LinkedHashSet;
-import java.util.Map;
 import java.util.Random;
 import java.util.Set;
+import java.util.concurrent.Executor;
+import java.util.concurrent.Executors;
 
 /**
  * Class for testing HandlersConfigurer.
@@ -138,16 +137,14 @@ public class HandlersConfigurerTestWrapper {
             protected void configure() {
                 // Needed by e.g. SearchHandler
                 bind(Linguistics.class).to(SimpleLinguistics.class).in(Scopes.SINGLETON);
-                bind(ContainerThreadPool.class).toInstance(
-                        new ContainerThreadPool(
-                                new ContainerThreadpoolConfig(new ContainerThreadpoolConfig.Builder()),
-                                new Metric() {
-                                    @Override public void set(String key, Number val, Context ctx) {}
-                                    @Override public void add(String key, Number val, Context ctx) {}
-                                    @Override public Context createContext(Map<String, ?> properties) { return null;}
-                                }));
+                bind(ContainerThreadPool.class).to(SimpleContainerThreadpool.class);
             }
         });
+    }
+
+    private static class SimpleContainerThreadpool implements ContainerThreadPool {
+        private final Executor executor = Executors.newCachedThreadPool();
+        @Override public Executor executor() { return executor; }
     }
 
 }
